# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Interview Backend API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:18080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/moderator/human-interviews/{interviewId}/complete": {
      "post": {
        "tags": [
          "moderator-candidate-application-controller"
        ],
        "operationId": "completeHumanInterview",
        "parameters": [
          {
            "name": "interviewId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanInterviewCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseHumanInterviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moderator/human-interviews/{interviewId}/cancel": {
      "post": {
        "tags": [
          "moderator-candidate-application-controller"
        ],
        "operationId": "cancelHumanInterview",
        "parameters": [
          {
            "name": "interviewId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseHumanInterviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moderator/companies/{companyId}/request-revision": {
      "post": {
        "tags": [
          "moderator-company-verification-controller"
        ],
        "operationId": "requestRevision",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCompanyVerificationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moderator/companies/{companyId}/reject": {
      "post": {
        "tags": [
          "moderator-company-verification-controller"
        ],
        "operationId": "reject",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCompanyVerificationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moderator/companies/{companyId}/approve": {
      "post": {
        "tags": [
          "moderator-company-verification-controller"
        ],
        "operationId": "approve",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCompanyVerificationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moderator/candidate-applications/{applicationId}/reject": {
      "post": {
        "tags": [
          "moderator-candidate-application-controller"
        ],
        "operationId": "reject_1",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCandidateApplicationReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moderator/candidate-applications/{applicationId}/human-interviews": {
      "post": {
        "tags": [
          "moderator-candidate-application-controller"
        ],
        "operationId": "scheduleHumanInterview",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanInterviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseHumanInterviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moderator/candidate-applications/{applicationId}/forward": {
      "post": {
        "tags": [
          "moderator-candidate-application-controller"
        ],
        "operationId": "forward",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCandidateApplicationReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moderator/candidate-applications/{applicationId}/approve": {
      "post": {
        "tags": [
          "moderator-candidate-application-controller"
        ],
        "operationId": "approve_1",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCandidateApplicationReviewResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/moderator/human-interviews/{interviewId}/reschedule": {
      "patch": {
        "tags": [
          "moderator-candidate-application-controller"
        ],
        "operationId": "rescheduleHumanInterview",
        "parameters": [
          {
            "name": "interviewId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanInterviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseHumanInterviewResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/moderator/companies": {
      "get": {
        "tags": [
          "moderator-company-verification-controller"
        ],
        "operationId": "getCompanies",
        "parameters": [
          {
            "name": "verificationStatus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING_VERIFICATION",
                "APPROVED",
                "REJECTED",
                "SUSPENDED"
              ]
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageModeratorCompanyListItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moderator/companies/{companyId}": {
      "get": {
        "tags": [
          "moderator-company-verification-controller"
        ],
        "operationId": "getCompany",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseModeratorCompanyDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moderator/candidate-applications": {
      "get": {
        "tags": [
          "moderator-candidate-application-controller"
        ],
        "operationId": "getReviewQueue",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "IN_REVIEW",
                "HUMAN_INTERVIEW_SCHEDULED",
                "DECISION_PENDING",
                "APPROVED",
                "REJECTED",
                "FORWARDED"
              ]
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageCandidateApplicationListItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/moderator/candidate-applications/{applicationId}": {
      "get": {
        "tags": [
          "moderator-candidate-application-controller"
        ],
        "operationId": "getReviewDetail",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCandidateApplicationDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyDocumentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "companyId": {
            "type": "integer",
            "format": "int64"
          },
          "uploadedByRecruiterProfileId": {
            "type": "integer",
            "format": "int64"
          },
          "documentType": {
            "type": "string"
          },
          "documentUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "PENDING",
              "SUSPENDED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "HumanInterviewCompleteRequest": {
        "type": "object",
        "properties": {
          "result": {
            "type": "string",
            "enum": [
              "PASSED",
              "FAILED",
              "NEEDS_REVIEW"
            ]
          },
          "note": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 0
          }
        },
        "required": [
          "result"
        ]
      },
      "ApiResponseHumanInterviewResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/HumanInterviewResponse"
          }
        }
      },
      "HumanInterviewResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "applicationId": {
            "type": "integer",
            "format": "int64"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "meetingUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PREPARING",
              "READY",
              "PENDING",
              "IN_PROGRESS",
              "COMPLETED",
              "FAILED",
              "CANCELLED"
            ]
          },
          "result": {
            "type": "string",
            "enum": [
              "PASSED",
              "FAILED",
              "NEEDS_REVIEW"
            ]
          },
          "note": {
            "type": "string"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DecisionRequest": {
        "type": "object",
        "properties": {
          "decisionNote": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 0
          }
        }
      },
      "ApiResponseCompanyVerificationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CompanyVerificationResponse"
          }
        }
      },
      "CompanyVerificationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "companyId": {
            "type": "integer",
            "format": "int64"
          },
          "moderatorProfileId": {
            "type": "integer",
            "format": "int64"
          },
          "decision": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REJECTED",
              "NEEDS_REVISION"
            ]
          },
          "note": {
            "type": "string"
          },
          "verifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiResponseCandidateApplicationReviewResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CandidateApplicationReviewResponse"
          }
        }
      },
      "CandidateApplicationReviewResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "reviewStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "IN_REVIEW",
              "HUMAN_INTERVIEW_SCHEDULED",
              "DECISION_PENDING",
              "APPROVED",
              "REJECTED",
              "FORWARDED"
            ]
          },
          "decisionNote": {
            "type": "string"
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time"
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time"
          },
          "forwardedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "HumanInterviewRequest": {
        "type": "object",
        "properties": {
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "meetingUrl": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "meetingUrl",
          "scheduledAt"
        ]
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
...
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponsePageModeratorCompanyListItemResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageModeratorCompanyListItemResponse"
          }
        }
      },
      "ModeratorCompanyListItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "recruiterProfileId": {
            "type": "integer",
            "format": "int64"
          },
          "industryId": {
            "type": "integer",
            "format": "int64"
          },
          "industryName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "websiteUrl": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "businessRegistrationNo": {
            "type": "string"
          },
          "verificationStatus": {
            "type": "string",
            "enum": [
              "PENDING_VERIFICATION",
              "APPROVED",
              "REJECTED",
              "SUSPENDED"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "PENDING",
              "SUSPENDED"
            ]
          }
        }
      },
      "PageModeratorCompanyListItemResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModeratorCompanyListItemResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "ApiResponseModeratorCompanyDetailResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ModeratorCompanyDetailResponse"
          }
        }
      },
      "ModeratorCompanyDetailResponse": {
        "type": "object",
        "properties": {
          "company": {
            "$ref": "#/components/schemas/CompanyResponse"
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyDocumentResponse"
            }
          },
          "verificationHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyVerificationResponse"
            }
          }
        }
      },
      "ApiResponsePageCandidateApplicationListItemResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageCandidateApplicationListItemResponse"
          }
        }
      },
...
      "PageCandidateApplicationListItemResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateApplicationListItemResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "ApiResponseCandidateApplicationDetailResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CandidateApplicationDetailResponse"
          }
        }
      },
      "CandidateApplicationDetailResponse": {
        "type": "object",
        "properties": {
          "application": {
            "$ref": "#/components/schemas/ApplicationSummaryResponse"
          },
          "candidate": {
            "$ref": "#/components/schemas/CandidateProfileResponse"
          },
          "submittedResume": {
            "$ref": "#/components/schemas/SubmittedResumeResponse"
          },
          "review": {
            "$ref": "#/components/schemas/CandidateApplicationReviewResponse"
          },
          "aiResult": {
            "$ref": "#/components/schemas/AiInterviewResultResponse"
          },
          "humanInterviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HumanInterviewResponse"
            }
          },
          "projectAssignments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectAssignmentSummaryResponse"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}